Paragraph styles based on ratio start
I am trying to make a set of paragraph styles that are based on the measure(line length) divided by 30. This is supposed to be an ideal length (but it probably silly). The measure is calculated as anything inside the margins in pts. Then divided to give an ideal type size for that measure. That is then set to as paragraph style Body. Then using that page shape based ratio new type sizes are created for different hierarchy need.